Issue #13: Make a deterministic wallet
Now with a master key that controls the whole wallet.

Unfortunately, this means that there won't be imports of any other addresses.
However, given the single master key, this will now deterministically make
wallets of every color_set. We now only need to store the master key and the
number of addresses that exist for each asset.

This is, of course, massively convenient. The master key can also make a
cold wallet for colored coins possible. The cold wallet would need the colored
coin definitions and how many addresses of each color exist in order to
make it maximally convenient.