Issue #66: Show transaction history for a color
Added a get_history_for_asset method in wallet_model which does most of the work.

Modified coloredcoinlib/blockchain.py to add a raw_address element to CTxOut.
This element is recorded from scriptPubKey which is in the raw transaction from bitcoincore.
This allows us to tie back transactions to addresses later on.

The actual getting history only shows the NET AFFECT on an address. So if an address
sent 1000 of color A coins and receives back 800 of color A coins, we'll only show that
color A sent 200 of color A coins.

Note I haven't implemented the scanhistory method. I don't think that's really necessary
since it's scan and then history of asset.