Issue #3: Added compose_tx_spec and compose_genesis_tx_spec
Also hooked it up all the way to the command-line.
I changed the txspec to take color_def instead of color_id.
I also changed the convention of using "0" to mark a target as uncolored.
We are now using a constant UNCOLORED_MARKER, similar to GENESIS_OUTPUT_MARKER
to signify that this is an uncolored target. This makes for a much more consistent
way of handling color definitions throughout.