Issue #75: color mempool transactions
Added a command-line command "addressbalance" which gives you the balance of each address instead of the sum of all the addresses of a particular color.
Added every send transaction to the utxo db, so the rest of the system can know about it.
After send, that command gets scanned right away.
The scan command will erase the mempool transactions in utxodb, which means they need to get rebuilt.