Issue #87: Make a method that takes a pycoin tx and converts to a composedTxSpec
Unfortunately, there's no way to do this in txspec.py without the coloredcoincontext. This is because Address depends on whether we're on testnet or not.
Similarly, the call to blockchain_state's get_tx is unavoidable. That is how we get the actual value of the utxo.
Even in pycoin's own scripts/send.py, it uses a call out to blockchain.info in order to get the actual amount that's available in the utxo.

I'm not exactly sure how to test it or in what context a call like this would happen. Please let me konw what I can do to do that.