Issue #145 Part 2:
Changed the output of coinlog to also have the colorvalues
(in the form of <colordef>: <colorvalue>) and the spending transactions
which is in the form of a list.

A little refactoring was done and a bug related to doing "ngccc.py scan"
when there are no new blocks was fixed.