Adding regtest support and logging support
REGTEST CHANGES
The regtest server url is taken from the config.json file of ngccc-
server. Regtest server is by default None, and if not explicitly changed
(by ngccc-server), regtest mode is not switched on.

Regtest status is implemented as a new 'testing_config' module, where a
'regtest_server' module-level variable stores whether regtest mode is
enabled.

Regtest mode is handled as a sub mode of the testnet mode. Those modules
that need it, imports the testing_config module and first checks if
we're in testnet mode and then checks if regtest_server is not None. If
so, certain tests in these modules' code are relaxed to allow regtest
mode, and switch to the regtest server url indicated.

CHANGES TO LOGGING
A new logger that logs to file is used, that does not segfault the
system: multiprocessing-logging. It is included as a dependency in
setup.py

The logger logs to a new "logs" directory and keeps the log file under a
fixed size. It can rotate logs but only one log is used right now. A log
format is used that is verbose: It includes line number, thread number,
process id and a number of other things. The idea is to use grep to
filter out what one wants to look at, filtering by thread, module, debug
level and so on. Critical messages and higher are also logged to stderr.

OTHER
There are PEP8 formatting changes in blockchain.py, since it was edited
(for regtest support). Removing one unused variable "height" in
blockchain.py too.

rpdb is included in setup.py as an optional dependency. It allows
debugging from a separate terminal window over a port, e.g. with netcat:
'netcat localhost 4444'.